First attempt at modularizing code/splitting into lib and tables mod.
Still need to split tables mod further to encapsulate difference between table functions and secret functions.
